requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.6
Markdown==2.6.9
psycopg2==2.7.3
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.10.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.44.0
